requirements.txt
playwright==1.52.0
openai==1.51.2
notion-client==2.3.0
python-dotenv==1.0.1
tenacity==9.0.0
